Split historyEngine.go into small files (#5972)
