+++ service/history/engine/engineimpl/remove_signal_mutable_state.go
@@ -1,0 +1,59 @@
+// Copyright (c) 2017-2021 Uber Technologies, Inc.
+// Portions of the Software are attributed to Copyright (c) 2021 Temporal Technologies Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package engineimpl
+
+import (
+	"context"
+
+	"github.com/uber/cadence/common/types"
+	"github.com/uber/cadence/service/history/execution"
+	"github.com/uber/cadence/service/history/workflow"
+)
+
+// RemoveSignalMutableState remove the signal request id in signal_requested for deduplicate
+func (e *historyEngineImpl) RemoveSignalMutableState(
+	ctx context.Context,
+	request *types.RemoveSignalMutableStateRequest,
+) error {
+
+	domainEntry, err := e.getActiveDomainByID(request.DomainUUID)
+	if err != nil {
+		return err
+	}
+	domainID := domainEntry.GetInfo().ID
+
+	workflowExecution := types.WorkflowExecution{
+		WorkflowID: request.WorkflowExecution.WorkflowID,
+		RunID:      request.WorkflowExecution.RunID,
+	}
+
+	return workflow.UpdateWithAction(ctx, e.executionCache, domainID, workflowExecution, false, e.timeSource.Now(),
+		func(wfContext execution.Context, mutableState execution.MutableState) error {
+			if !mutableState.IsWorkflowExecutionRunning() {
+				return workflow.ErrNotExists
+			}
+
+			mutableState.DeleteSignalRequested(request.GetRequestID())
+
+			return nil
+		})
+}
