+++ service/history/engine/engineimpl/reset_sticky_tasklist.go
@@ -1,0 +1,64 @@
+// Copyright (c) 2017-2021 Uber Technologies, Inc.
+// Portions of the Software are attributed to Copyright (c) 2021 Temporal Technologies Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package engineimpl
+
+import (
+	"context"
+
+	"github.com/uber/cadence/common"
+	"github.com/uber/cadence/common/types"
+	"github.com/uber/cadence/service/history/execution"
+	"github.com/uber/cadence/service/history/workflow"
+)
+
+// ResetStickyTaskList reset the volatile information in mutable state of a given types.
+// Volatile information are the information related to client, such as:
+// 1. StickyTaskList
+// 2. StickyScheduleToStartTimeout
+// 3. ClientLibraryVersion
+// 4. ClientFeatureVersion
+// 5. ClientImpl
+func (e *historyEngineImpl) ResetStickyTaskList(
+	ctx context.Context,
+	resetRequest *types.HistoryResetStickyTaskListRequest,
+) (*types.HistoryResetStickyTaskListResponse, error) {
+
+	if err := common.ValidateDomainUUID(resetRequest.DomainUUID); err != nil {
+		return nil, err
+	}
+	domainID := resetRequest.DomainUUID
+
+	err := workflow.UpdateWithAction(ctx, e.executionCache, domainID, *resetRequest.Execution, false, e.timeSource.Now(),
+		func(wfContext execution.Context, mutableState execution.MutableState) error {
+			if !mutableState.IsWorkflowExecutionRunning() {
+				return workflow.ErrAlreadyCompleted
+			}
+			mutableState.ClearStickyness()
+			return nil
+		},
+	)
+
+	if err != nil {
+		return nil, err
+	}
+	return &types.HistoryResetStickyTaskListResponse{}, nil
+}
