+++ service/history/engine/engineimpl/respond_activity_task_canceled.go
@@ -1,0 +1,121 @@
+// Copyright (c) 2017-2021 Uber Technologies, Inc.
+// Portions of the Software are attributed to Copyright (c) 2021 Temporal Technologies Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package engineimpl
+
+import (
+	"context"
+	"time"
+
+	"github.com/uber/cadence/common"
+	"github.com/uber/cadence/common/log/tag"
+	"github.com/uber/cadence/common/metrics"
+	"github.com/uber/cadence/common/types"
+	"github.com/uber/cadence/service/history/execution"
+	"github.com/uber/cadence/service/history/workflow"
+)
+
+// RespondActivityTaskCanceled completes an activity task failure.
+func (e *historyEngineImpl) RespondActivityTaskCanceled(
+	ctx context.Context,
+	req *types.HistoryRespondActivityTaskCanceledRequest,
+) error {
+
+	domainEntry, err := e.getActiveDomainByID(req.DomainUUID)
+	if err != nil {
+		return err
+	}
+	domainID := domainEntry.GetInfo().ID
+	domainName := domainEntry.GetInfo().Name
+
+	request := req.CancelRequest
+	token, err0 := e.tokenSerializer.Deserialize(request.TaskToken)
+	if err0 != nil {
+		return workflow.ErrDeserializingToken
+	}
+
+	workflowExecution := types.WorkflowExecution{
+		WorkflowID: token.WorkflowID,
+		RunID:      token.RunID,
+	}
+
+	var activityStartedTime time.Time
+	var taskList string
+	err = workflow.UpdateWithAction(ctx, e.executionCache, domainID, workflowExecution, true, e.timeSource.Now(),
+		func(wfContext execution.Context, mutableState execution.MutableState) error {
+			if !mutableState.IsWorkflowExecutionRunning() {
+				return workflow.ErrAlreadyCompleted
+			}
+
+			scheduleID := token.ScheduleID
+			if scheduleID == common.EmptyEventID { // client call CompleteActivityById, so get scheduleID by activityID
+				scheduleID, err0 = getScheduleID(token.ActivityID, mutableState)
+				if err0 != nil {
+					return err0
+				}
+			}
+			ai, isRunning := mutableState.GetActivityInfo(scheduleID)
+
+			// First check to see if cache needs to be refreshed as we could potentially have stale workflow execution in
+			// some extreme cassandra failure cases.
+			if !isRunning && scheduleID >= mutableState.GetNextEventID() {
+				e.metricsClient.IncCounter(metrics.HistoryRespondActivityTaskCanceledScope, metrics.StaleMutableStateCounter)
+				e.logger.Error("Encounter stale mutable state in RecordActivityTaskCanceled",
+					tag.WorkflowDomainName(domainName),
+					tag.WorkflowID(workflowExecution.GetWorkflowID()),
+					tag.WorkflowRunID(workflowExecution.GetRunID()),
+					tag.WorkflowScheduleID(scheduleID),
+					tag.WorkflowNextEventID(mutableState.GetNextEventID()),
+				)
+				return workflow.ErrStaleState
+			}
+
+			if !isRunning || ai.StartedID == common.EmptyEventID ||
+				(token.ScheduleID != common.EmptyEventID && token.ScheduleAttempt != int64(ai.Attempt)) {
+				return workflow.ErrActivityTaskNotFound
+			}
+
+			if _, err := mutableState.AddActivityTaskCanceledEvent(
+				scheduleID,
+				ai.StartedID,
+				ai.CancelRequestID,
+				request.Details,
+				request.Identity); err != nil {
+				// Unable to add ActivityTaskCanceled event to history
+				return &types.InternalServiceError{Message: "Unable to add ActivityTaskCanceled event to history."}
+			}
+
+			activityStartedTime = ai.StartedTime
+			taskList = ai.TaskList
+			return nil
+		})
+	if err == nil && !activityStartedTime.IsZero() {
+		scope := e.metricsClient.Scope(metrics.HistoryClientRespondActivityTaskCanceledScope).
+			Tagged(
+				metrics.DomainTag(domainName),
+				metrics.WorkflowTypeTag(token.WorkflowType),
+				metrics.ActivityTypeTag(token.ActivityType),
+				metrics.TaskListTag(taskList),
+			)
+		scope.RecordTimer(metrics.ActivityE2ELatency, time.Since(activityStartedTime))
+	}
+	return err
+}
