+++ service/history/engine/engineimpl/respond_activity_task_failed.go
@@ -1,0 +1,141 @@
+// Copyright (c) 2017-2021 Uber Technologies, Inc.
+// Portions of the Software are attributed to Copyright (c) 2021 Temporal Technologies Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package engineimpl
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/uber/cadence/common"
+	"github.com/uber/cadence/common/log/tag"
+	"github.com/uber/cadence/common/metrics"
+	"github.com/uber/cadence/common/types"
+	"github.com/uber/cadence/service/history/execution"
+	"github.com/uber/cadence/service/history/workflow"
+)
+
+// RespondActivityTaskFailed completes an activity task failure.
+func (e *historyEngineImpl) RespondActivityTaskFailed(
+	ctx context.Context,
+	req *types.HistoryRespondActivityTaskFailedRequest,
+) error {
+
+	domainEntry, err := e.getActiveDomainByID(req.DomainUUID)
+	if err != nil {
+		return err
+	}
+	domainID := domainEntry.GetInfo().ID
+	domainName := domainEntry.GetInfo().Name
+
+	request := req.FailedRequest
+	token, err0 := e.tokenSerializer.Deserialize(request.TaskToken)
+	if err0 != nil {
+		return workflow.ErrDeserializingToken
+	}
+
+	workflowExecution := types.WorkflowExecution{
+		WorkflowID: token.WorkflowID,
+		RunID:      token.RunID,
+	}
+
+	var activityStartedTime time.Time
+	var taskList string
+	err = workflow.UpdateWithActionFunc(
+		ctx,
+		e.executionCache,
+		domainID,
+		workflowExecution,
+		e.timeSource.Now(),
+		func(wfContext execution.Context, mutableState execution.MutableState) (*workflow.UpdateAction, error) {
+			if !mutableState.IsWorkflowExecutionRunning() {
+				return nil, workflow.ErrAlreadyCompleted
+			}
+
+			scheduleID := token.ScheduleID
+			if scheduleID == common.EmptyEventID { // client call CompleteActivityById, so get scheduleID by activityID
+				scheduleID, err0 = getScheduleID(token.ActivityID, mutableState)
+				if err0 != nil {
+					return nil, err0
+				}
+			}
+			ai, isRunning := mutableState.GetActivityInfo(scheduleID)
+
+			// First check to see if cache needs to be refreshed as we could potentially have stale workflow execution in
+			// some extreme cassandra failure cases.
+			if !isRunning && scheduleID >= mutableState.GetNextEventID() {
+				e.metricsClient.IncCounter(metrics.HistoryRespondActivityTaskFailedScope, metrics.StaleMutableStateCounter)
+				e.logger.Error("Encounter stale mutable state in RecordActivityTaskFailed",
+					tag.WorkflowDomainName(domainName),
+					tag.WorkflowID(workflowExecution.GetWorkflowID()),
+					tag.WorkflowRunID(workflowExecution.GetRunID()),
+					tag.WorkflowScheduleID(scheduleID),
+					tag.WorkflowNextEventID(mutableState.GetNextEventID()),
+				)
+				return nil, workflow.ErrStaleState
+			}
+
+			if !isRunning || ai.StartedID == common.EmptyEventID ||
+				(token.ScheduleID != common.EmptyEventID && token.ScheduleAttempt != int64(ai.Attempt)) {
+				e.logger.Warn(fmt.Sprintf(
+					"Encounter non existing activity in RecordActivityTaskFailed: isRunning: %t, ai: %#v, token: %#v.",
+					isRunning, ai, token),
+					tag.WorkflowDomainName(domainName),
+					tag.WorkflowID(workflowExecution.GetWorkflowID()),
+					tag.WorkflowRunID(workflowExecution.GetRunID()),
+					tag.WorkflowScheduleID(scheduleID),
+					tag.WorkflowNextEventID(mutableState.GetNextEventID()),
+				)
+				return nil, workflow.ErrActivityTaskNotFound
+			}
+
+			postActions := &workflow.UpdateAction{}
+			ok, err := mutableState.RetryActivity(ai, req.FailedRequest.GetReason(), req.FailedRequest.GetDetails())
+			if err != nil {
+				return nil, err
+			}
+			if !ok {
+				// no more retry, and we want to record the failure event
+				if _, err := mutableState.AddActivityTaskFailedEvent(scheduleID, ai.StartedID, request); err != nil {
+					// Unable to add ActivityTaskFailed event to history
+					return nil, &types.InternalServiceError{Message: "Unable to add ActivityTaskFailed event to history."}
+				}
+				postActions.CreateDecision = true
+			}
+
+			activityStartedTime = ai.StartedTime
+			taskList = ai.TaskList
+			return postActions, nil
+		},
+	)
+	if err == nil && !activityStartedTime.IsZero() {
+		scope := e.metricsClient.Scope(metrics.HistoryRespondActivityTaskFailedScope).
+			Tagged(
+				metrics.DomainTag(domainName),
+				metrics.WorkflowTypeTag(token.WorkflowType),
+				metrics.ActivityTypeTag(token.ActivityType),
+				metrics.TaskListTag(taskList),
+			)
+		scope.RecordTimer(metrics.ActivityE2ELatency, time.Since(activityStartedTime))
+	}
+	return err
+}
